Frontend components reorganized in folders
