+++ frontend/package.json
@@ -5,6 +5,9 @@
   "dependencies": {
     "@date-io/core": "^2.5.0",
     "@date-io/date-fns": "^1.3.13",
+    "@devexpress/dx-react-core": "^2.6.2",
+    "@devexpress/dx-react-scheduler": "^2.6.2",
+    "@devexpress/dx-react-scheduler-material-ui": "^2.6.2",
     "@material-ui/core": "^4.9.9",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/pickers": "^3.2.10",
@@ -13,8 +16,12 @@
     "@testing-library/user-event": "^7.2.1",
     "axios": "^0.19.2",
     "date-fns": "^2.12.0",
+    "dnd-core": "^10.0.2",
     "match-sorter": "^4.1.0",
     "react": "^16.13.1",
+    "react-big-scheduler": "^0.2.6",
+    "react-dnd": "^10.0.2",
+    "react-dnd-html5-backend": "^10.0.2",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.4.1",
