+++ frontend/src/components/ClinicAdminComponents/RequestedAppointmentsComponent.js
@@ -58,8 +58,8 @@ export default function RequestedAppointmentsComponent() {
   return (
     <div>
       {requests.length===0 ? <h1>No appointment requests</h1> : null}
-      {requests.map((request) => (
-        <AppointmentRequestRow appointment={request} rooms={rooms} removeRequest={() => {removeRequest(request.id)}}/>
+      {requests.map((request, index) => (
+        <AppointmentRequestRow key={index} index={index} appointment={request} rooms={rooms} removeRequest={() => {removeRequest(request.id)}}/>
       ))}
       
     </div>
