+++ frontend/src/components/HomePageLayout.js
@@ -15,19 +15,19 @@ import ListItem from '@material-ui/core/ListItem';
 import ListItemText from '@material-ui/core/ListItemText';
 import IconButton from '@material-ui/core/IconButton';
 import AccountCircle from '@material-ui/icons/AccountCircle';
-// import MenuItem from '@material-ui/core/MenuItem';
-// import Menu from '@material-ui/core/Menu';
+import MenuItem from '@material-ui/core/MenuItem';
+import Menu from '@material-ui/core/Menu';
 import { useHistory } from 'react-router-dom';
 import Collapse from '@material-ui/core/Collapse';
 import ExpandLess from '@material-ui/icons/ExpandLess';
 import ExpandMore from '@material-ui/icons/ExpandMore';
-import ClinicStaffList from './ClinicStaffList';
-import AppointmentTypesComponent from './AppointmentTypesComponent';
-import OneClickAppointmentComponent from './OneClickAppointmentComponent';
-import MakeAppointmentDoctorComponent from './MakeAppointmentDoctorComponent';
-import RequestedAppointmentsComponent from './RequestedAppointmentsComponent';
+import ClinicStaffList from './StaffComponents/ClinicStaffList';
+import AppointmentTypesComponent from './ClinicAdminComponents/AppointmentTypesComponent';
+import OneClickAppointmentComponent from './ClinicAdminComponents/OneClickAppointmentComponent';
+import MakeAppointmentDoctorComponent from './DoctorComponents/MakeAppointmentDoctorComponent';
+import RequestedAppointmentsComponent from './ClinicAdminComponents/RequestedAppointmentsComponent';
 import UserProfileComponent from './UserProfileComponent';
-import ClinicRoomComponent from './ClinicRoomComponent';
+import ClinicRoomComponent from './ClinicAdminComponents/ClinicRoomComponent';
 
 const drawerWidth = 240;
 
@@ -106,8 +106,8 @@ const useStyles = makeStyles((theme) => ({
 
 export default function ClippedDrawer() {
   const classes = useStyles();
-  // const [anchorEl, setAnchorEl] = React.useState(null);
-  // const open = Boolean(anchorEl);
+  const [anchorEl, setAnchorEl] = React.useState(null);
+  const open = Boolean(anchorEl);
   const [openMenu, setOpen] = React.useState(false);
 
   const history = useHistory();
@@ -116,13 +116,13 @@ export default function ClippedDrawer() {
     setOpen(!openMenu);
   };
 
-  // const handleMenu = (event) => {
-  //   setAnchorEl(event.currentTarget);
-  // };
+  const handleMenu = (event) => {
+    setAnchorEl(event.currentTarget);
+  };
 
-  // const handleClose = () => {
-  //   setAnchorEl(null);
-  // };
+  const handleClose = () => {
+    setAnchorEl(null);
+  };
 
   function goTo(url) {
     history.push(url);
@@ -147,7 +147,7 @@ export default function ClippedDrawer() {
                 <AccountCircle />
                 My account
               </IconButton>
-              {/* <Menu
+              <Menu
                 id="menu-appbar"
                 anchorEl={anchorEl}
                 anchorOrigin={{
@@ -163,7 +163,7 @@ export default function ClippedDrawer() {
                 onClose={handleClose}
               >
                 <MenuItem onClick={() => {goTo('/acc')}}>Profile</MenuItem>
-              </Menu> */}
+              </Menu>
             </div>
         </Toolbar>
       </AppBar>
@@ -245,7 +245,7 @@ export default function ClippedDrawer() {
             <UserProfileComponent />
           </Route>
           <Route exact path="/">
-            <Home />
+            <Home  />
           </Route>
           <Route>
             * 404 NOT FOUND *
