+++ frontend/src/components/StaffComponents/AddDoctorComponent.js
